+++ tests-integration/src/test/java/com/contentful/vaultintegration/AllTheThingsTest.java
@@ -20,6 +20,7 @@ import com.contentful.vault.Asset;
 import com.contentful.vault.Vault;
 import com.contentful.vaultintegration.lib.allthethings.AllTheThingsResource;
 import com.contentful.vaultintegration.lib.allthethings.AllTheThingsSpace;
+
 import org.junit.Test;
 import org.robolectric.RuntimeEnvironment;
 import org.robolectric.annotation.Config;
@@ -28,7 +29,7 @@ import static com.contentful.vault.BaseFields.REMOTE_ID;
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.fail;
 
-@Config(manifest = "src/main/AndroidManifest.xml")
+@Config(manifest = "src/main/AndroidManifest.xml", sdk = 23)
 public class AllTheThingsTest extends BaseTest {
   @Override protected void setupVault() {
     vault = Vault.with(RuntimeEnvironment.application, AllTheThingsSpace.class);