+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Version [3.1.0] - 2018-02-13
+## Version [3.2.0] - TBD
+- Change: use https for asset urls.
+- Change: Use CDA SDK 10.0.0 and throw if environments are used.
+- Change: Locale handling: CDA 10 uses `/locales` and not `/space.locales`, update tests and internal structure.
+
+## Version [3.1.1] - 2018-02-13
 - Add: update CDA SDK
 - Remove: Synthetic methods
 
@@ -123,7 +128,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Version 0.9.0 - 2015-05-18
 Initial release.
  
-[unreleased]: https://github.com/contentful/vault/compare/vault-parent-3.0.1...HEAD
+[unreleased]: https://github.com/contentful/vault/compare/vault-parent-3.2.0...HEAD
+[3.2.0]: https://github.com/contentful/vault/compare/vault-parent-3.1.1...vault-parent-3.2.0
+[3.1.1]: https://github.com/contentful/vault/compare/vault-parent-3.1.0...vault-parent-3.1.1
+[3.1.0]: https://github.com/contentful/vault/compare/vault-parent-3.0.2...vault-parent-3.1.0
+[3.0.2]: https://github.com/contentful/vault/compare/vault-parent-3.0.1...vault-parent-3.0.2
 [3.0.1]: https://github.com/contentful/vault/compare/vault-parent-3.0.0...vault-parent-3.0.1
 [3.0.0]: https://github.com/contentful/vault/compare/vault-parent-2.1.0...vault-parent-3.0.0
 [2.1.0]: https://github.com/contentful/vault/compare/vault-parent-2.0.0...vault-parent-2.1.0